+++ get_next_line.h
@@ -6,7 +6,7 @@
 /*   By: tlucille <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/01/06 12:51:42 by tlucille          #+#    #+#             */
-/*   Updated: 2020/01/09 12:29:56 by tlucille         ###   ########.fr       */
+/*   Updated: 2020/01/13 10:22:49 by tlucille         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,7 +24,7 @@
 char		*gnl_strchr(const char *s, int c);
 char		*gnl_strnew(size_t size);
 char		*gnl_strjoin(char **s1, char *s2, int fr, int ret);
-char		*gnl_strdup(const char *s1, int empty);
+char		*gnl_strdup(const char *s1, ssize_t empty);
 void		*gnl_memset(void *b, int c, size_t len);
 int			get_next_line(int fd, char **line);
 size_t		gnl_strlen(const char *str);
